Cuboid size int instead of double

+++ src/utils/Generator.h
@@ -19,7 +19,7 @@ namespace Generator {
      * @param mass
      * @param typeId
      */
-    void cuboid(ParticleContainer &container, std::array<double, 3> position, std::array<double, 3> size, double meshWidth,
+    void cuboid(ParticleContainer &container, std::array<double, 3> position, std::array<int, 3> size, double meshWidth,
            std::array<double, 3> velocity, double mass, int typeId);
 
     /**
